Update screen format a bit and pktgen.isSending change
The Lua pktgen.isSending has been changed to return 'n' or 'y'
instead of zero or non-zero.

The rfc2544.lua script is not a bit closer to working, but still
has some problems. The big problem is the X520 NICs are not
returning the correct number of packets sent. The stats for opackets
comes from the hardware registers.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>