tcp/udp: set ipv4 version_ihl so checksum is valid
Both protocols call rte_ipv4_udptcp_cksum() to build a L4 checksum. That
function calls rte_ipv4_hdr_len() to get the header length and subtracts
that value from the L3 length to determine the L4 length. It looks like
this code was last modified in October 2020.

I believe the existing code was not setting the ipv4 ihl field before
calling rte_ipv4_udptcp_cksum(). On my system with default settings,
the version_ihl byte is 0x6f which means the header length is 60 bytes,
the L3 length is 46, so the L4 length is -14. Thus no checksum is
computed.

With this change, I see valid checksums on 64 byte packets using TCP and
UDP under IPv4.