Fix loss of data stored in udata64 mbuf field
DPDK v20.11 removed mbuf's udata64 field, and changed type of the
dynfield1 to uint32_t.

Due to pktgen's define for DPDK v20.11+:
  lib/common/pg_compat.h:#define udata64                 dynfield1[0]
udata64 field was being mapped to the first entry of the dynfield1
array.

Because of that, accessing dynfield1 directly was causing 2 issues:
  * As dynfield1 is 32-bit array, using it interchangeably with the
    udata64 is causing only the first 32-bits of the intended 64-bits to
    be copied
  * dynfield1 may also be used by the other parts of the DPDK and it may
    cause conflicts

To fix that, the dynfield1 is no longer mapped to the udata64 as those
fields are uncompatible. Instead, the dynamic field API is used to get
the pktgen data offset from the dynfield1 array.

It was wrapped into an getter function, which could be reworked if
backward compatibility will be needed.

Signed-off-by: Michal Krawczyk <mk@semihalf.com>
Reviewed-by: Igor Chauskin <igorch@amazon.com>