pcap: use pcap max packet len to size mempool
The pcap module creates a memory pool where each element is sized to
hold a jumbo frame. Therefore the memory requirement is dependent on the
number of packets in the pcap and not the size of the packets.

This change determines the largest packet in the pcap and sizes the
mempool elements based on that value. This can dramatically increase the
number of packets that be loaded (dependent on the size of the packets
of course).