Account for delay when computing next tx cycle
Using curr_tsc to compute the next tx cycle causes the dealy between the
expected tx cycle and curr_tsc to be lost, producing a lower rate than
the desired one. Computing the next tx cycle from the old value of
tx_next_cycle prevents this problem and allows to obtain a rate closer
to the expected one.

Signed-off-by: Federico Parola <federico.parola@polito.it>