Fix hanging when quit early
There is a race condition in setting pktgen cores' 'stop' flags.
When 'quit' is entered immediately after pktgen is launched it is
possible for 'stop' flag to be cleared by rx/tx threads after it is set
in the main function, leading to pktgen hanging indefinitely.

Instead of increasing the 'Wait for the lcores to start up' 250 ms delay
this commit introduces another quick and dirty fix: a started-stopped
pair of flags. This of course assumes that cores are not restarted in
the process' lifetime, which is true for current pktgen version.

Plus this commit gets rid of a couple now-redundant initializations.

Signed-off-by: Philipp <phila775@gmail.com>