fix rx mtu on cards with higher theoretical MTU
some cards have theoretical MTU >60000, which doesn't work with Ethernet.
Now pktgen uses 1500 as default rx MTU and a maximum of 9600, when jumbo
frames are enabled. The calculation of the maximum theoretical MTU was
also adapted by substracting the overhead according to an example
implementation of DPDK (examples/l3fwd-graph/main.c).
Also enabled scatter offloading when available, which is needed on mlnx
cards.

Signed-off-by: dlinder <19273958+linderd@users.noreply.github.com>